refactor(core): extract maintenance/dedup-keys.ts — plan+apply dedup-key backfill
